requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.4
openai>=1.0.0
python-dotenv>=1.0.0
selenium>=4.0.0
requests>=2.31.0
